chat_backend/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
# backports.zoneinfo==0.2.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.7
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
filelock==3.9.0
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
mpmath==1.2.1
networkx==3.0
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
regex==2023.8.8
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sqlparse==0.4.3
sympy==1.11.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
# torch==2.0.1+cu117
# torchaudio==2.0.2+cu117
# torchvision==0.15.2+cu117
tqdm==4.66.1
typing-extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.17.0
